requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
accelerate>=0.20.0
Pillow>=9.0.0
requests>=2.28.0 